w08d02 TodoSQL: implement 'delete Todo' feature

+++ week-08/day-02/workshop/todosql/src/main/java/com/greenfoxacademy/todosql/repository/TodoRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Component;
 
 @Component
-public interface TodoRepository extends CrudRepository<Todo, Long> {
+public interface TodoRepository extends CrudRepository<Todo, Integer> {
 
   List<Todo> findByIsDoneFalse();
